# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/calendars/ics": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Imports an ICS file via JSON body or file upload. Calendar name is extracted from ICS properties (X-WR-CALNAME) or falls back to \"Untitled Calendar\"",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "Import ICS File",
                "parameters": [
                    {
                        "description": "Import ICS request (JSON) - calendar_name is optional",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/calendar.ImportICSRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Calendar name override (optional - will use ICS properties if not provided)",
                        "name": "calendar_name",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "ICS file to upload (required for file upload)",
                        "name": "ics_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ICS file imported successfully",
                        "schema": {
                            "$ref": "#/definitions/calendar.ImportICSResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid request body or ICS data",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication required",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/users/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the authenticated user's profile information from JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Profile",
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/model.UserProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication required",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - User not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "calendar.ImportICSRequest": {
            "type": "object",
            "required": [
                "ics_data"
            ],
            "properties": {
                "calendar_name": {
                    "type": "string"
                },
                "ics_data": {
                    "type": "string"
                }
            }
        },
        "calendar.ImportICSResponse": {
            "type": "object",
            "properties": {
                "calendar": {
                    "$ref": "#/definitions/model.Calendar"
                },
                "events_count": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "model.Account": {
            "description": "OAuth account information",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "description": "e.g. example@gmail.com",
                    "type": "string"
                },
                "expiry": {
                    "description": "Access token expiry",
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "provider": {
                    "description": "e.g. \"google\", \"github\"",
                    "type": "string"
                },
                "provider_id": {
                    "description": "e.g. Google sub, GitHub ID",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
...
        "model.Calendar": {
            "description": "Calendar",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "event_color": {
                    "type": "string"
                },
                "event_nickname": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "source": {
                    "$ref": "#/definitions/model.CalendarSource"
                },
                "source_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "synced_at": {
                    "type": "string"
                },
                "time_zone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "0"
                },
                "visibility": {
                    "$ref": "#/definitions/model.CalendarVisibility"
                }
            }
        },
...
        "model.CalendarSource": {
            "type": "string",
            "enum": [
                "google",
                "isc"
            ],
            "x-enum-varnames": [
                "SourceGoogle",
                "SourceISC"
            ]
        },
        "model.CalendarVisibility": {
            "type": "string",
            "enum": [
                "public",
                "private"
            ],
            "x-enum-varnames": [
                "CalendarVisibilityPublic",
                "CalendarVisibilityPrivate"
            ]
        },
...
        "model.ErrorResponse": {
            "description": "Error response structure",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Detailed error information",
                    "type": "string",
                    "example": "invalid email or password"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Authentication failed"
                },
                "success": {
                    "description": "Always false for error responses",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "model.User": {
            "description": "User account information",
            "type": "object",
            "properties": {
                "accounts": {
                    "description": "Associated OAuth accounts",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Account"
                    }
                },
                "created_at": {
                    "description": "Account creation timestamp",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "display_name": {
                    "description": "User's display name",
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "description": "Unique user identifier",
                    "type": "string",
                    "example": "123456789"
                },
                "password": {
                    "description": "Password hash (excluded from responses)",
                    "type": "string"
                },
                "picture": {
                    "description": "Profile picture URL",
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "updated_at": {
                    "description": "Last update timestamp",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "model.UserProfileResponse": {
            "description": "User profile response",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User profile retrieved successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        }
    }
}